refact : divide the home page in component
